+++ apps/components/src/App.tsx
@@ -15,6 +15,7 @@ import PopoverPage from './popover/page';
 import BasePage from './base/page';
 import FilterPanelBasePage from './filter-panel-base/page';
 import ChipPage from './chip/page';
+import AlertPage from './alert/page';
 import TooltipPage from './tooltip/page';
 import IconsPage from './icons/page';
 import SearchInputPage from './search-input/page';
@@ -54,6 +55,7 @@ export default function App() {
         <Route path="/base" element={<BasePage />} />
         <Route path="/filter-panel-base" element={<FilterPanelBasePage />} />
         <Route path="/chip" element={<ChipPage />} />
+        <Route path="/alert" element={<AlertPage />} />
         <Route path="/tooltip" element={<TooltipPage />} />
         <Route path="/table" element={<TablePage />} />
         <Route path="/tabs" element={<TabsPage />} />